# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "demo",
        "description": " This is the first demo project",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Meresa",
            "email": "meresa@gmail.com"
        },
        "license": {
            "name": "Demo platform",
            "url": "http://localhost:3000/api/comment"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/product/{productId}": {
            "get": {
                "description": "single product display ",
                "summary": "get product ",
                "operationId": "product",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "Id",
                        "default": "60464badb501e14018153d8d",
                        "description": "Enter product Id",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get product by id"
                    }
                },
                "tags": [
                    "product"
                ]
            }
        },
...
        "/news/{newsId}":{
            "get": {
                "summary": "display news",
                "operationId": "get all",
                "tags": ["news"],
                "description": "returns news info",
                "consumes":[
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "news",
                        "type": "string",
                        "description": "enter news id",
                        "default":"60a1152e75566c33fc674c7f",
                        "required": true
                    }
                ],
                "responses": {
                    "200":{
                        "description": "display all news"
                    }
                }
            }
        }
    },
...
}